- Moved DateFormatter extension to APICategory - Added some protocols for graphQLFilter from QueryPredicate - Moved pretty print json serialization to test package
